backend/requirements.txt
fastapi
uvicorn
boto3
pydantic
python-dotenv
python-jose[cryptography]
requests
sqlalchemy
pymysql
